requirements.txt
click==8.0.1
colorama==0.4.4
contextlib2==21.6.0
Flask==2.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
protobuf==3.17.3
python-dotenv==0.19.0
python-http-client==3.3.2
schema==0.7.4
sendgrid==6.8.0
six==1.16.0
starkbank-ecdsa==1.1.1
Werkzeug==2.0.1